requirements.txt
python-telegram-bot==20.7
requests
python-dotenv